requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

# Optional feature groups
telethon>=1.34.0   # required for autopost & telegram send
python-telegram-bot>=20.0,<21.0  # required for interactive bot
feedparser>=6.0,<7.0  # required for news digest
